+++ framework-api-jmx/src/main/java/uk.gov.justice.services.jmx/CatchupMBean.java
@@ -4,5 +4,6 @@ import javax.management.MXBean;
 
 @MXBean
 public interface CatchupMBean {
+
     void doCatchupRequested();
 }