+++ framework-api-jmx/src/main/java/uk.gov.justice.services.jmx/RebuildMBean.java
@@ -1,0 +1,6 @@
+package uk.gov.justice.services.jmx;
+
+public interface RebuildMBean {
+
+    void doRebuildRequested();
+}